move some clz from autorest to client
