+++ src/KubernetesClient/Authentication/ExecTokenProvider.cs
@@ -1,4 +1,3 @@
-using k8s.Autorest;
 using k8s.KubeConfigModels;
 using System.Net.Http.Headers;
 using System.Threading;
