+++ src/KubernetesClient/Authentication/ServiceClientCredentials.cs
@@ -2,8 +2,10 @@
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
 
-namespace k8s.Autorest
+namespace k8s.Authentication
 {
     /// <summary>
     /// ServiceClientCredentials is the abstraction for credentials used by ServiceClients accessing REST services.
