+++ src/KubernetesClient/Authentication/TokenCredentials.cs
@@ -2,8 +2,10 @@
 // Licensed under the MIT License. See License.txt in the project root for license information.
 
 using System.Net.Http;
+using System.Threading;
+using System.Threading.Tasks;
 
-namespace k8s.Autorest
+namespace k8s.Authentication
 {
     /// <summary>
     /// Token based credentials for use with a REST Service Client.
