+++ src/KubernetesClient/KubernetesClientConfiguration.cs
@@ -1,4 +1,4 @@
-using k8s.Autorest;
+using k8s.Authentication;
 using System.Net.Http;
 using System.Security.Cryptography.X509Certificates;
